LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-js-expose-expected.txt

Text

PASS Element.getBoundingClientRect() returns the actual rendered location
PASS Range.getBoundingClientRect() returns the actual rendered location
FAIL Element.offset* return adjusted offsets assert_equals: expected 200 but got 1500
PASS IntersectionObserver should fire when anchored element is moved into visible area
